introduce default element in some message content
When sending messages with performatives accept_proposal, agree, cfp, failure, not_understood, propose, refuse, and reject_proposal, the required content would be a tuple ending with propositional values (e.g., to express reasons, or preconditions).
Now these values can be omitted, and the compiler automatically replaces them with the propositional value True.
Example:
Both

send message agree (A, P) to R

and

send message agree A to R

where A is an action, P is a proposition and R is a receiver, are now valid syntaxes. In the second case, the new compiler converts the content of the message into the tuple (A, True).